refactor functiout out, return map version in results
